storybook: parametrize the 13 near-duplicate story smoke tests
test_stories_basic.py had 13 copy-pasted ~6-line functions doing the
same thing per story. Replace with a single parametrized test driven
by the real registered names (all_story_names()), plus a separate
"every expected story is registered" assertion against a hardcoded
name list so a story that's implemented but never imported into
stories/__init__.py still fails loudly. LineStyleIcons and
FontFamilyOptions keep their own tests since they assert something
different (empty controls list, populated widget).
